requirements.txt
numpy==1.22.2
matplotlib==3.5.2
PyYAML==6.0
GitPython==3.1.27
scipy==1.8.1
tqdm==4.64.0
secure-smtplib==0.1.1
